moved Card function to be its own component
